added more unit tests

+++ src/NavisNFT.sol
@@ -10,16 +10,14 @@ import "@openzeppelin/contracts/token/ERC1155/extensions/ERC1155Supply.sol";
 import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 import "@openzeppelin/contracts/utils/Counters.sol";
 
-
-
 contract NavisNFT is ERC1155, AccessControl, ERC1155Pausable, ERC1155Burnable, ERC1155Supply {
     using Counters for Counters.Counter;
-    Counters.Counter public  _tokenIdTracker; 
+
+    Counters.Counter public _tokenIdTracker;
     uint256 public constant PREMIUM_ID_OFFSET = 6; // Offsets premium ID from free ID to ensure Premium is non-fungible.
 
     //
-    
-    mapping(uint=>string[]) nftAbilities;
+    mapping(uint256 => string[]) nftAbilities;
 
     bytes32 public constant URI_SETTER_ROLE = keccak256("URI_SETTER_ROLE");
     bytes32 public constant PAUSER_ROLE = keccak256("PAUSER_ROLE");
@@ -27,32 +25,30 @@ contract NavisNFT is ERC1155, AccessControl, ERC1155Pausable, ERC1155Burnable, E
     string public name;
     string public symbol;
     //mapping (address => )
-    mapping (address=>bool) userHasMinted;
+    mapping(address => bool) userHasMinted;
 
     // Mapping from ship ID to ship abilities
     mapping(uint256 => string[]) public shipAbilities;
-    mapping (uint256=> uint256) public premiumShipID;
-    mapping (address=> uint256[]) public userToNFT;
-    mapping(uint256=>uint256) public premiumShipIDToType;
+    mapping(uint256 => uint256) public premiumShipID;
+    mapping(address => uint256[]) public userToNFT;
+    mapping(uint256 => uint256) public premiumShipIDToType;
 
     struct premiumShipData {
-        uint256 id; 
-        string[] shipAbilities;}
+        uint256 id;
+        string[] shipAbilities;
+    }
 
     premiumShipData[] public premiumShipAbilities; // Keeps a record of the ship abilities mapped to the id.
 
     address public feeCollector; // Address that collects the token fees
     IERC20 public navisToken;
-    
-    uint256 public constant MINT_PRICE = 20 * 10**18; 
+
+    uint256 public constant MINT_PRICE = 20 * 10 ** 18;
 
     // Mapping from ship type code to URI
     mapping(uint256 => string) public shipTypeURIs;
 
-
-
-
-   constructor(address _navisTokenAddress) ERC1155("") {
+    constructor(address _navisTokenAddress) ERC1155("") {
         _grantRole(DEFAULT_ADMIN_ROLE, msg.sender);
         _grantRole(PAUSER_ROLE, msg.sender);
         _grantRole(MINTER_ROLE, msg.sender);
@@ -63,31 +59,28 @@ contract NavisNFT is ERC1155, AccessControl, ERC1155Pausable, ERC1155Burnable, E
 
         // Initialize URIs for each ship type, assuming 75 types
         for (uint256 i = 1; i <= 75; i++) {
-            shipTypeURIs[i] = string(abi.encodePacked("https://gnfd-testnet-sp1.bnbchain.org/view/navis-nft-test/", uint2str(i), ".json"));
+            shipTypeURIs[i] = string(
+                abi.encodePacked("https://gnfd-testnet-sp1.bnbchain.org/view/navis-nft-test/", uint2str(i), ".json")
+            );
         }
     }
-
 
     function setURI(string memory newuri) public onlyRole(URI_SETTER_ROLE) {
         _setURI(newuri);
     }
 
-
     function pause() public onlyRole(PAUSER_ROLE) {
         _pause();
     }
-
 
     function unpause() public onlyRole(PAUSER_ROLE) {
         _unpause();
     }
 
-
     function setFeeCollector(address _feeCollector) public onlyRole(DEFAULT_ADMIN_ROLE) {
         feeCollector = _feeCollector;
     }
 
-    
     // This is not gas efficient. How to make it gas efficient?
     // Function to update ship abilities
     function updateShipAbilities(uint256 _id, string[] memory _abilities) public onlyRole(MINTER_ROLE) {
@@ -102,42 +95,40 @@ contract NavisNFT is ERC1155, AccessControl, ERC1155Pausable, ERC1155Burnable, E
         return shipAbilities[_id];
     }
 
-//@notice This mints fungible tokens
-function mintFree() public {
-    require(!userHasMinted[msg.sender], "User already minted free NFT");
-    userHasMinted[msg.sender] = true;
-    for (uint256 i = 0; i < 6; i++) {
-        _mint(msg.sender, i, 1, "");
-        userToNFT[msg.sender].push(i);  
-        _setURI(shipTypeURIs[i]); 
-    }        
-}
+    //@notice This mints fungible tokens
+    function mintFree() public {
+        require(!userHasMinted[msg.sender], "User already minted free NFT");
+        userHasMinted[msg.sender] = true;
+        for (uint256 i = 0; i < 6; i++) {
+            _mint(msg.sender, i, 1, "");
+            userToNFT[msg.sender].push(i);
+            _setURI(shipTypeURIs[i]);
+        }
+    }
 
-   
-//@notice This mints non-fungible tokens
-function mintPremium(uint256 shipType) public returns (uint256) {
-    require(navisToken.transferFrom(msg.sender, feeCollector, MINT_PRICE), "Fee transfer failed");
-    require(shipType > 5 && shipType <= 75, "Invalid ship type");
-    uint256 newTokenId = _tokenIdTracker.current() + PREMIUM_ID_OFFSET;
-    _mint(msg.sender, newTokenId, 1, "");
-    userToNFT[msg.sender].push(newTokenId); 
-    premiumShipIDToType[newTokenId] = shipType; 
-    _setURI(shipTypeURIs[shipType]); 
-    _tokenIdTracker.increment(); 
+    //@notice This mints non-fungible tokens
+    function mintPremium(uint256 shipType) public returns (uint256) {
+        require(navisToken.transferFrom(msg.sender, feeCollector, MINT_PRICE), "Fee transfer failed");
+        require(shipType > 5 && shipType <= 75, "Invalid ship type");
+        uint256 newTokenId = _tokenIdTracker.current() + PREMIUM_ID_OFFSET;
+        _mint(msg.sender, newTokenId, 1, "");
+        userToNFT[msg.sender].push(newTokenId);
+        premiumShipIDToType[newTokenId] = shipType;
+        _setURI(shipTypeURIs[shipType]);
+        _tokenIdTracker.increment();
 
-    return  newTokenId;
-}
+        return newTokenId;
+    }
 
     function getUserShipIDs(address _user) public view returns (uint256[] memory) {
-    return userToNFT[_user];
-}
-
+        return userToNFT[_user];
+    }
 
     function setShipTypeURI(uint256 shipType, string memory uri) public onlyRole(URI_SETTER_ROLE) {
         require(shipType > 0 && shipType <= 50, "Invalid ship type");
         shipTypeURIs[shipType] = uri;
     }
-   
+
     // Helper function to convert uint256 to string
     function uint2str(uint256 _i) internal pure returns (string memory _uintAsString) {
         if (_i == 0) {
@@ -152,7 +143,7 @@ function mintPremium(uint256 shipType) public returns (uint256) {
         bytes memory bstr = new bytes(len);
         uint256 k = len;
         while (_i != 0) {
-            k = k-1;
+            k = k - 1;
             uint8 temp = (48 + uint8(_i - _i / 10 * 10));
             bytes1 b1 = bytes1(temp);
             bstr[k] = b1;
@@ -170,12 +161,7 @@ function mintPremium(uint256 shipType) public returns (uint256) {
         super._update(from, to, ids, values);
     }
 
-    function supportsInterface(bytes4 interfaceId)
-        public
-        view
-        override(ERC1155, AccessControl)
-        returns (bool)
-    {
+    function supportsInterface(bytes4 interfaceId) public view override(ERC1155, AccessControl) returns (bool) {
         return super.supportsInterface(interfaceId);
     }
 }
